Remove second rebase
Because we are always rebasing on top of the latest with DLRN, the
second rebase is not required at this point. By running only the first
rebase, we end up in the desired state with the downstream patches
standing on top of the latest upstream commit, and duplicate commits
now carried upstream being dropped.