Add -patches branch when it does not exist
When the -patches branch is not present in the repository, we may want
to create it automatically. If that is the case, we will create the
branch based on the commit received, then push it.

This commit introduces a new setting in patch_rebaser.ini, to enable
the feature if needed. It also honors the dev_mode setting, to avoid
pushing during development.