Use the PATCHES_BRANCH env var instead of guessing
When finding the patches branch for a repository, we used to try
to guess it. However, in the DLRN use case, we have a PATCHES_BRANCH
environment variable we can use, so use it if available.

We keep guessing optionally, if the environment variable is missing,
so we can use patch_rebaser in a different context.