Push rebase results when dropping downstream-only patches
Sometimes, a patch is proactively backported downstream ahead of
merging upstream. When that happens, the rebase results do not get
pushed because there are no functional changes between the branches as
reported by the cherry command.

However, we still want to push in this case to make sure the latest
commit hash is the same between upstream and downstream for consistency
and because other tooling relies on that. This patch adds a check to
confirm the latest upstream commit we rebased on is present on the
patches remote - if it isn't, then we push regardless of the cherry
results.

We still don't want to push all the time to avoid unnecessary churn
(e.g. hashes for d/s patches changing after each rebase) and
meaningless tags being pushed to the remote.