Update github action versions.
Update supported versions of python, as we only went up to python 3.9,
and 3.6 is now EOL:
https://devguide.python.org/versions/

Release versions of checkout and python actions also needed to be
updated due to github changing the node they run on:
https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/

Removing mergify because github now offers functionality that is 'close
enough', and the churn in mergify config has turned out to be more
effort than we feel it is worth.

Minor tweaks to test dependencies to keep in line with our other
projects.

Removed incorrect version references.

Minor fix to tox.in to whitelist external commands, and
structure it more consostently with other projects.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>