Update remote if url has changed (#34)
Previously we were checking if the remote name is there to add it.
Now, if the patches_repo url has changed, but we maintain the same
remote_name, this is updated.

For doing this, we just check if the remote_name exists and if the
patches_repo differs from its value. If so, we just remove the remote
so next line below it'd recreate it.

Not it's based on new git_wrapper version that allows to return the
remote dict with its value but also to remove the remote.

Requires: https://github.com/release-depot/git_wrapper/pull/114